(wip) disabled webpack optimizer + fixed contextMenus
